+++ sync-for-commerce/src/codatsynccommerce/models/operations/get_sync_status.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import syncstatus as shared_syncstatus
-from typing import Any, Optional
+from ...models.shared import syncstatus as shared_syncstatus
+from typing import Optional
 
 
 @dataclasses.dataclass
@@ -19,14 +19,10 @@ class GetSyncStatusRequest:
 class GetSyncStatusResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    bad_request: Optional[Any] = dataclasses.field(default=None)
-    r"""Bad Request"""
-    not_found: Optional[Any] = dataclasses.field(default=None)
-    r"""Not Found"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
     sync_status: Optional[shared_syncstatus.SyncStatus] = dataclasses.field(default=None)
     r"""Success"""