+++ sync-for-commerce/src/codatsynccommerce/models/operations/update_config_text_sync_flow.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import localization as shared_localization
+from ...models.shared import localization as shared_localization
 from typing import Dict, Optional
 
 
@@ -11,11 +11,11 @@ from typing import Dict, Optional
 class UpdateConfigTextSyncFlowResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     localization_info: Optional[Dict[str, shared_localization.Localization]] = dataclasses.field(default=None)
     r"""Success"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""